requirements.txt
Flask
SQLAlchemy
pandas
bcrypt
PyJWT
psycopg2
